# swagger.json
{
    "openapi": "3.0.3",
    "info": {
      "title": "Postgre CRUD",
      "description": "Testing postgre crud",
      "version": "0.0.1"
    },
    "paths": {
        "/api/product": {
            "get": {
                "description": "List all products",
                "parameters": [
                    {
                        "in": "query",
                        "name": "limit",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "in": "query",
                        "name": "offset",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200":{
                        "description": "una lista de usuarios",
                        "content": {
                            "application/json":{
                                "schema" : {
                                    "type": "object",
                                    "properties" : {
                                        "data": {
                                            "type" : "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "leche"
                                                    }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "post a new product",
                "parameters": [
                    {
                        "in": "body",
                        "name": "",
                        "schema": {
                            "$ref": "#/definitions/Product"
                        }
                    }
                ]
            },
            "delete": {
                "description": "delete a product by id",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ]
            },
            "put": {
                "description": "update a product",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200":{
                        "description": "el usuario actualizado",
                        "content": {
                            "application/json":{
                                "schema" : {
                                    "type": "object",
                                    "properties" : {
                                        "data": {
                                            "type" : "object",
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "leche"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions":{
        "Product":{
            "type":"object",
            "required":[
                "name"
            ],
            "properties":{
                "name":{
                    "type":"string"
                },
                "description":{
                    "type":"string"
                },
                "price":{
                    "type":"integer"
                }
            }
        }
    }
  }